feat : h2database to postrepository, userrepository
h2database는 init및 처리만 Post model 생성
