+++ src/main/java/db/PostRepository.java
@@ -1,0 +1,4 @@
+package db;
+
+public class PostRepository {
+}
