+++ src/main/java/model/Post.java
@@ -1,0 +1,26 @@
+package model;
+
+public class Post {
+    private String writer;
+    private String title;
+    private String content;
+
+    public String getWriter() {
+        return writer;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public String getContent() {
+        return content;
+    }
+
+
+    public Post(String writer, String title, String content) {
+        this.writer = writer;
+        this.title = title;
+        this.content = content;
+    }
+}
