+++ src/main/java/webserver/WebServer.java
@@ -5,6 +5,8 @@ import java.net.Socket;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
+import db.Database;
+import db.H2Database;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -20,6 +22,8 @@ public class WebServer {
         } else {
             port = Integer.parseInt(args[0]);
         }
+
+        H2Database.initTable();
 
         ExecutorService executorService = Executors.newFixedThreadPool(THREAD_POOL_SIZE);
